Colorado/cross_validation/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cross-validation"
version = "1.1.0"
description = "Agente de validación cruzada KYB para empresas mexicanas"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.29",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27"]
portales = [
    "playwright>=1.49",
    "openpyxl>=3.1",
    "Pillow>=10.0",
    "pytesseract>=0.3",
]

[project.scripts]
cross-validation = "cross_validation.main:cli"

[tool.setuptools.packages.find]
include = ["cross_validation*"]
